fix: rename licence spelling
