+++ app/services/licences/fetch-licence-summary.service.js
