+++ app/services/licences/view-licence-summary.service.js
@@ -6,7 +6,7 @@
  */
 
 const FetchLicenceAbstractionConditionsService = require('./fetch-licence-abstraction-conditions.service.js')
-const FetchLicenceSummaryService = require('./fetch-license-summary.service')
+const FetchLicenceSummaryService = require('./fetch-licence-summary.service')
 const ViewLicenceSummaryPresenter = require('../../presenters/licences/view-licence-summary.presenter')
 const ViewLicenceService = require('./view-licence.service')
