+++ test/services/licences/view-licence-summary.service.test.js
@@ -13,7 +13,7 @@ const LicenceModel = require('../../../app/models/licence.model.js')
 
 // Things we need to stub
 const FetchLicenceAbstractionConditionsService = require('../../../app/services/licences/fetch-licence-abstraction-conditions.service.js')
-const FetchLicenceSummaryService = require('../../../app/services/licences/fetch-license-summary.service')
+const FetchLicenceSummaryService = require('../../../app/services/licences/fetch-licence-summary.service')
 const ViewLicenceService = require('../../../app/services/licences/view-licence.service')
 // Thing under test
 const ViewLicenceSummaryService = require('../../../app/services/licences/view-licence-summary.service')
@@ -29,7 +29,7 @@ describe('View Licence service summary', () => {
       purposeIds: [],
       numberOfConditions: 0
     })
-    Sinon.stub(ViewLicenceService, 'go').resolves({ licenceName: 'fake license' })
+    Sinon.stub(ViewLicenceService, 'go').resolves({ licenceName: 'fake licence' })
   })
 
   afterEach(() => {
@@ -64,7 +64,7 @@ describe('View Licence service summary', () => {
           endDate: null,
           id: '2c80bd22-a005-4cf4-a2a2-73812a9861de',
           licenceHolder: 'Unregistered licence',
-          licenceName: 'fake license',
+          licenceName: 'fake licence',
           monitoringStations: [],
           purposes: null,
           region: 'South West',
